bug: Move `listArtifacts()` from `ToolContext` to the parent `CallbackContext` class
